+++ frontend/src/App.tsx
@@ -1,8 +1,9 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Home from './pages/Home';
+import ImageDescription from './pages/ImageDescription';
 import UploadPage from './pages/audiofeature';
 import ContentGrid from './pages/contentGrid';
 import Search from './pages/search';
+import Home from './pages/Home';
 import './App.css';
 
 function App() {
@@ -10,6 +11,7 @@ function App() {
     <Router>
       <Routes>
         <Route path="/" element={<Home />} />
+        <Route path="/describe" element={<ImageDescription />} />
         <Route path="/upload" element={<UploadPage />} />
         <Route path="/view" element={<ContentGrid />} />
         <Route path='/search' element={<Search />} />
